Implement first-blush of encapsulation interface.
Conversion to wire protocol bytes is stubbed out until I'm back to the
ol' homestead and can look to see how other clients handle this section.

Additionally, geos are a bit distinct in that they are two values-- the
NewGeoValue signature may change to be better compatible with
vl.Object() once I'm more than 0% familiar with the codebase.